app/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
mutagen==1.46.0
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
gunicorn==20.1.0
